@@ -2,62 +2,75 @@ package com.lisakzbigniew.flashcardsapi.model;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
 
-/** A collection of flash cards belonging to
- * an owner(might support mulitiple users later on) */
+/**
+ * A collection of flash cards belonging to
+ * an owner(might support mulitiple users later on)
+ */
 @Entity
 public class CardCollection {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
+
     @Column(nullable = false)
     private String name;
-    @OneToMany List<Card> cards;
+
+    @OneToMany
+    List<Card> cards;
+
     @Column(nullable = false, columnDefinition = "varchar(255) default 'root'")
     private String owner;
-
 
     public Long getId() {
         return id;
     }
+
     public void setId(Long id) {
         this.id = id;
     }
+
     public String getName() {
         return name;
     }
+
     public void setName(String name) {
         this.name = name;
     }
+
     public List<Card> getCards() {
         return cards;
     }
+
     public void setCards(List<Card> cards) {
         this.cards = cards;
     }
+
     public String getOwner() {
         return owner;
     }
+
     public void setOwner(String owner) {
         this.owner = owner;
     }
 
     @Override
     public String toString() {
-        return "{" +
-            " id='" + getId() + "'" +
-            ", name='" + getName() + "'" +
-            ", cards='" + getCards() + "'" +
-            ", owner='" + getOwner() + "'" +
-            "}";
+        return "{\n" +
+                "id = " + getId().toString() + ", \n" +
+                "name = " + getName().toString() + ", \n" +
+                "cards = [\n" + getCards().stream().map(Card::toString).collect(Collectors.joining(",\n")) + "\n], \n" +
+                "owner = " + getOwner().toString() + " \n" +
+                "}\n";
     }
 
     @Override
@@ -68,10 +81,10 @@ public class CardCollection {
             return false;
         }
         CardCollection cardCollection = (CardCollection) o;
-        return Objects.equals(id, cardCollection.id) &&
-        Objects.equals(name, cardCollection.name) &&
-        Objects.equals(cards, cardCollection.cards) &&
-        Objects.equals(owner, cardCollection.owner);
+        return Objects.equals(id, cardCollection.id) ||
+                Objects.equals(name, cardCollection.name) &&
+                Objects.equals(cards, cardCollection.cards) &&
+                Objects.equals(owner, cardCollection.owner);
     }
 
     @Override
@@ -79,5 +92,4 @@ public class CardCollection {
         return Objects.hash(id, name, cards, owner);
     }
 
-    
 }
