@@ -2,62 +2,74 @@ package com.lisakzbigniew.flashcardsapi.model;
 
 import java.util.Objects;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 
-/**A Phrase in a given language*/
+/** A Phrase in a given language */
 @Entity
 public class Phrase {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
+
     private String content;
+
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
     private Language lang;
+
     public Long getId() {
         return id;
     }
+
     public void setId(Long id) {
         this.id = id;
     }
+
     public String getContent() {
         return content;
     }
+
     public void setContent(String content) {
         this.content = content;
     }
+
     public Language getLang() {
         return lang;
     }
+
     public void setLang(Language lang) {
         this.lang = lang;
     }
 
+    public boolean inLang(Language lang){
+        return this.lang.equals(lang);
+    }
+
     @Override
     public String toString() {
-        return "{" +
-            " content='" + getContent() + "'" +
-            ", lang='" + getLang() + "'" +
-            "}";
+        return "{\n" +
+                "content = " + getContent().toString() + ", \n" +
+                "lang = " + getLang().toString() + " \n" +
+                "} \n";
     }
-    
 
     @Override
     public boolean equals(Object o) {
         if (o == this)
             return true;
-        if (!(o instanceof Phrase)) {
+        if (!(o instanceof Phrase phrase)) {
             return false;
         }
-        Phrase phrase = (Phrase) o;
-        return Objects.equals(id, phrase.id) && Objects.equals(content, phrase.content) && Objects.equals(lang, phrase.lang);
+        return Objects.equals(id, phrase.id) ||
+                Objects.equals(content, phrase.content) &&
+                        Objects.equals(lang, phrase.lang);
     }
 
     @Override
@@ -65,5 +77,4 @@ public class Phrase {
         return Objects.hash(id, content, lang);
     }
 
-    
 }
